requirements.txt
click==8.0.3
hydra-core==1.1.1
matplotlib==3.5.1
numpy==1.21.5
python-dotenv==0.19.2
setuptools==60.2.0
torch==1.10.1
